refactor: deployment helpers
