+++ src/deploy/helpers/configuratorModificator.ts
