+++ test/helpers/deploymentTokenInfra.ts
@@ -1,5 +1,5 @@
-import { FaucetToken, SimplePriceFeed, Fauceteer } from "../../../build/types";
-import { DeploymentManager } from '../../../plugins/deployment_manager';
+import { FaucetToken, SimplePriceFeed, Fauceteer } from '../../build/types';
+import { DeploymentManager } from '../../plugins/deployment_manager';
 
 // Helper function to create tokens
 export async function makeToken(
@@ -8,13 +8,6 @@ export async function makeToken(
   name: string,
   decimals: number
 ): Promise<FaucetToken> {
-  // Check if token already exists
-  const existingToken = await deploymentManager.contract(symbol);
-  if (existingToken) {
-    return existingToken as FaucetToken;
-  }
-  
-  // Deploy new token
   const mint = (BigInt(1000000) * 10n ** BigInt(decimals)).toString();
   return deploymentManager.deploy(symbol, 'test/FaucetToken.sol', [mint, name, decimals, symbol]);
 }
@@ -26,13 +19,6 @@ export async function makePriceFeed(
   initialPrice: number,
   decimals: number
 ): Promise<SimplePriceFeed> {
-  // Check if price feed already exists
-  const existingPriceFeed = await deploymentManager.contract(alias);
-  if (existingPriceFeed) {
-    return existingPriceFeed as SimplePriceFeed;
-  }
-  
-  // Deploy new price feed
   return deploymentManager.deploy(alias, 'test/SimplePriceFeed.sol', [initialPrice * decimals, decimals]);
 }
 
@@ -40,12 +26,5 @@ export async function makePriceFeed(
 export async function makeFauceteer(
   deploymentManager: DeploymentManager
 ): Promise<Fauceteer> {
-  // Check if fauceteer already exists
-  const existingFauceteer = await deploymentManager.contract('fauceteer');
-  if (existingFauceteer) {
-    return existingFauceteer as Fauceteer;
-  }
-  
-  // Deploy new fauceteer
   return deploymentManager.deploy('fauceteer', 'test/Fauceteer.sol', []);
 }
